maint: remove rest api support for diagrams
