+++ doc/source/user_guide/api/create-conn.rst
@@ -16,7 +16,7 @@ Ansys SysML2 API Connector
 PySAM SysML2 uses the standard API to load and publish information in a model.
 
 .. literalinclude:: ../../_static/code/download-diagrams.py
-    :lines: 35-40
+    :lines: 34-39
     :language: python
     :caption: Connect to the SysML2 API server
 
@@ -31,19 +31,18 @@ The ``SysML2APIConnector`` class is typically used with project managers for hig
 * ``execute_query(project_id, query)``: Sends a query (in JSON format) to the standard API.
 * ``create_commit(project_id, commit)``: Sends a commit (in JSON format) to the standard API.
 
-SAM REST API Connector
-----------------------
+SAM API Connector
+-----------------
 
-If you need to access diagrams and their associated elements (available only on Ansys servers), you can use the :class:`SamRestApiConnector <ansys.sam.sysml2.diagrams.api.sam_rest_api_connector.SamRestApiConnector>` connector.
+If you need to access diagrams and their associated elements (available only on Ansys servers), you can use the :class:`SamApiConnector <ansys.sam.sysml2.diagrams.api.sam_api_connector.SamApiConnector>` connector.
 
 .. literalinclude:: ../../_static/code/download-diagrams.py
-    :lines: 42-46
+    :lines: 41-45
     :language: python
-    :caption: Connect to the SAM REST API server
+    :caption: Connect to the SAM API server
 
 The ``SamApiConnector`` class provides access to diagrams and diagram-related data (Ansys servers only) through the following methods:
 
-* ``get_project_data(model_id)``: Returns project data from the SAM API using the project ID.
 * ``get_diagrams_info(project_id)``: Returns metadata and information for all diagrams within a specific project.
 * ``get_single_diagram_info(project_id, diagram_id)``: Returns detailed information for a single diagram within a project.
 * ``get_diagram_image_as_svg(project_id, diagram_id)``: Downloads a diagram rendered as SVG format.
