+++ src/ansys/sam/sysml2/diagrams/__init__.py
@@ -21,6 +21,5 @@
 # SOFTWARE.
 """Diagram module for PySAM SysML2 diagram library."""
 
-from .api import SamRestApiConnector  # noqa as we export name
-from .sam_diagram_manager import SAMDiagramManager  # noqa as we export name
+from .api import SamApiConnector  # noqa as we export name
 from .tools import SamDiagramDownloader  # noqa as we export name
