+++ src/ansys/sam/sysml2/diagrams/api/__init__.py
@@ -23,4 +23,3 @@
 """API package for the SAM API connector."""
 
 from .sam_api_connector import SamApiConnector  # noqa as we export name
-from .sam_rest_api_connector import SamRestApiConnector  # noqa as we export name
